E2E tests + some code fixes

+++ crates/client/src/clients/producer/future_record.rs
@@ -9,7 +9,7 @@ pub struct FutureRecord {
     pub value: Vec<u8>,
     pub topic: String,
     /// Set only if you want to force a specific partition, skipping partitioner mechanism
-    pub partition: Option<i16>,
+    pub partition: Option<i32>,
     pub headers: HashMap<String, Vec<u8>>,
 }
